Address requested changes
